making order out of the mess of the repository and adding the file for euler project problem#14
