--- Hello World.py
@@ -1,3 +1,0 @@
-print('Hello World!')
-print(2**8)
-input()
