--- input_test.py
@@ -1,5 +1,0 @@
-while True:
-	reply = input('Enter text: ')
-	if reply == 'stop': break
-	print(reply.upper())
-
