+++ math/3B1B_1_v2.py
