+++ tests/Hello World.py
@@ -1,0 +1,3 @@
+print('Hello World!')
+print(2**8)
+input()
