+++ tests/input_test.py
@@ -1,0 +1,5 @@
+while True:
+	reply = input('Enter text: ')
+	if reply == 'stop': break
+	print(reply.upper())
+
