+++ tests/script1.py
@@ -1,0 +1,6 @@
+import sys
+print(sys.platform)
+print(2**100)
+x='Spam!'
+print(x*8)
+input()
